+++ inst/essais/KleinFibonacci.R
@@ -1,0 +1,37 @@
+library(RcppColors)
+library(jacobi)
+
+# the modified Cayley transformation
+Phi <- function(z) (1i*z + 1) / (z + 1i)
+PhiInv <- function(z) {
+  1i + (2i*z) / (1i - z)
+}
+# background color
+bkgcol <- "#ffffff"
+# make the color mapping
+f <- function(x, y) {
+  z <- complex(real = x, imaginary = y)
+  w <- PhiInv(z)
+  ifelse(
+    Mod(z) > 0.96, 
+    NA_complex_,
+    ifelse(
+      y < 0, 
+      -1/w, w
+    )
+  )
+}
+x <- seq(-1, 1, length.out = 2048)
+y <- seq(-1, 1, length.out = 2048)
+Z <- outer(x, y, f)
+K <- kleinj(Z) / 1728
+G <- K / (1 - K - K*K)
+image <- colorMap7(G, bkgcolor = bkgcol)
+# plot
+opar <- par(mar = c(0,0,0,0), bg = bkgcol)
+plot(
+  c(-1, 1), c(-1, 1), type = "n", 
+  xlab = NA, ylab = NA, axes = FALSE, asp = 1
+)
+rasterImage(image, -1, -1, 1, 1)
+par(opar)